Added a destroy route for saved links Created the destroy method in the links controller
How it works is it will first see how many folders have the specific link you are looking to destroy in, since links in the database are not duplicated. If there is only one folder with the link that will be destroyed, the link is destroyed. If there is more than one folder that has the link, the folderlink is destroyed and the link is left alone.